# serializer: re-use marshaller buffers when possible