--- a/pkg/serializer/internal/metrics/iterable_series.go
+++ b/pkg/serializer/internal/metrics/iterable_series.go
@@ -112,6 +112,7 @@ func (series *IterableSeries) MarshalSplitCompress(bufferContext *marshaler.Buff
 	var err error
 	var compressor *stream.Compressor
 	buf := bufferContext.PrecompressionBuf
+	buf.Reset()
 	ps := molecule.NewProtoStream(buf)
 	payloads := transaction.BytesPayloads{}