@scottopell 2024-01-18
Contrary to my expectations, the SMP experiments show no gain here, and local microbenchmarks show a small _decrease_ in performance.
I will not merge this.
```
                                            │ bench_main.txt │       bench_reuse_buffers.txt        │
                                            │     sec/op     │    sec/op     vs base                │
Series/000005-items/05-points/10-tags/pb-36     288.2µ ± 31%   334.3µ ± 15%        ~ (p=0.052 n=10)
Series/000005-items/05-points/50-tags/pb-36     344.2µ ±  3%   369.1µ ± 11%   +7.25% (p=0.000 n=10)
Series/000005-items/10-points/10-tags/pb-36     321.4µ ±  2%   619.1µ ± 57%  +92.65% (p=0.000 n=10)
Series/000005-items/10-points/50-tags/pb-36     376.4µ ±  3%   462.5µ ± 24%  +22.90% (p=0.000 n=10)
Series/000010-items/05-points/10-tags/pb-36     309.2µ ±  1%   326.6µ ±  5%   +5.65% (p=0.000 n=10)
Series/000010-items/05-points/50-tags/pb-36     368.0µ ±  3%   384.6µ ±  2%   +4.52% (p=0.000 n=10)
Series/000010-items/10-points/10-tags/pb-36     375.4µ ±  2%   391.6µ ±  3%   +4.33% (p=0.001 n=10)
Series/000010-items/10-points/50-tags/pb-36     433.5µ ±  2%   478.3µ ±  1%  +10.35% (p=0.000 n=10)
Series/000100-items/05-points/10-tags/pb-36     775.4µ ±  1%   774.4µ ±  3%        ~ (p=0.684 n=10)
Series/000100-items/05-points/50-tags/pb-36     814.9µ ±  1%   778.6µ ±  0%   -4.45% (p=0.002 n=10)
Series/000100-items/10-points/10-tags/pb-36     1.229m ±  3%   1.288m ±  6%   +4.78% (p=0.000 n=10)
Series/000100-items/10-points/50-tags/pb-36     1.282m ±  1%   1.364m ±  4%   +6.39% (p=0.000 n=10)
Series/000500-items/05-points/10-tags/pb-36     2.473m ±  0%   2.572m ±  8%   +4.00% (p=0.000 n=10)
Series/000500-items/05-points/50-tags/pb-36     2.502m ±  1%   2.496m ±  3%        ~ (p=0.481 n=10)
Series/000500-items/10-points/10-tags/pb-36     4.935m ±  0%   4.925m ±  0%   -0.20% (p=0.000 n=10)
Series/000500-items/10-points/50-tags/pb-36     4.971m ±  0%   4.962m ±  0%   -0.20% (p=0.000 n=10)
Series/001000-items/05-points/10-tags/pb-36     4.585m ±  0%   4.591m ±  0%        ~ (p=0.165 n=10)
Series/001000-items/05-points/50-tags/pb-36     4.616m ±  0%   4.625m ±  0%   +0.19% (p=0.009 n=10)
Series/001000-items/10-points/10-tags/pb-36     9.650m ±  0%   9.616m ±  0%   -0.35% (p=0.000 n=10)
Series/001000-items/10-points/50-tags/pb-36     9.698m ±  0%   9.689m ±  0%        ~ (p=0.190 n=10)
Series/010000-items/05-points/10-tags/pb-36     54.51m ±  0%   54.48m ±  0%        ~ (p=0.971 n=10)
Series/010000-items/05-points/50-tags/pb-36     54.55m ±  0%   55.58m ±  2%   +1.89% (p=0.002 n=10)
Series/010000-items/10-points/10-tags/pb-36     111.4m ±  0%   111.4m ±  3%        ~ (p=0.912 n=10)
Series/010000-items/10-points/50-tags/pb-36     111.4m ±  0%   111.4m ±  0%        ~ (p=0.684 n=10)
Series/100000-items/05-points/10-tags/pb-36     31.50m ±  0%   30.64m ±  3%   -2.70% (p=0.005 n=10)
Series/100000-items/05-points/50-tags/pb-36     31.50m ±  0%   30.66m ±  1%   -2.68% (p=0.000 n=10)
Series/100000-items/10-points/10-tags/pb-36     63.09m ±  0%   61.53m ±  0%   -2.47% (p=0.000 n=10)
Series/100000-items/10-points/50-tags/pb-36     63.14m ±  0%   61.64m ±  1%   -2.37% (p=0.002 n=10)
geomean                                         3.724m         3.904m         +4.86%
```